Refactoring runsCount for main question

+++ src/main/java/homeworks/Questionable.java
@@ -4,5 +4,5 @@ package homeworks;
  * todo use this interface in previous homeworks
  */
 public interface Questionable {
-    long getAnswer();
+    int getAnswer();
 }
